+++ examples/erp/genseki/collections/users.ts
@@ -13,7 +13,15 @@ export const fields = builder.fields('user', (fb) => ({
   }),
   image: fb.columns('image', {
     type: 'media',
-    mimeTypes: ['image/jpeg', 'image/png', 'image/webp'],
+    label: 'Image',
+    isRequired: true,
+    uploadOptions: {
+      maxSize: 1024 * 1024 * 2, // 2MB
+      limit: 1,
+      accept: 'image/*',
+      mimeTypes: ['image/jpeg', 'image/png', 'image/webp'],
+      pathName: 'dev-uploads/test', // storage name
+    },
   }),
 }))
 
@@ -24,6 +32,8 @@ const list = builder.list(fields, {
     sortBy: ['name'],
   },
 })
+
+const update = builder.update(fields, {})
 
 const api = {
   example: builder.endpoint(
@@ -51,4 +61,5 @@ export const usersCollection = builder.collection({
   slug: 'users',
   list: list,
   api: api,
+  update: update,
 })