+++ packages/react/src/core/file-storage-adapters/handlers/grab-delete-obj-signed-url.ts
@@ -4,31 +4,37 @@ import type { AnyContextable } from '../../context'
 import { createEndpoint } from '../../endpoint'
 import type { StorageAdapter } from '../generic-adapter'
 
-export function grabPermanentObject<const TContext extends AnyContextable>(
+export function grabDeleteObjUrl<const TContext extends AnyContextable>(
   context: TContext,
   uploadAdapter?: StorageAdapter
 ) {
   return createEndpoint(
     context,
     {
-      method: 'GET',
-      path: '/storage/permanent-obj',
+      method: 'DELETE',
+      path: '/storage/delete-obj-signed-url',
       query: z.object({
         key: z.string(),
       }),
       responses: {
         200: z.object({
           message: z.string(),
+          signedUrl: z.string(),
         }),
       },
     },
     async (args) => {
-      if (!uploadAdapter) throw new Error('Storage adpater is missing at server configuration')
+      if (!uploadAdapter) throw new Error('Storage adapter is missing at server configuration')
+
+      const { message, data } = await uploadAdapter.generateDeleteObjectSignedUrl({
+        key: args.query.key,
+      })
 
       return {
         status: 200,
         body: {
-          message: 'This route is for future use',
+          message: message,
+          signedUrl: data.deleteObjectUrl,
         },
       }
     }