+++ packages/react/src/core/file-storage-adapters/s3/upload-adapter.ts
@@ -1,4 +1,5 @@
 import {
+  DeleteObjectCommand,
   GetObjectCommand,
   PutObjectCommand,
   S3Client,
@@ -6,21 +7,42 @@ import {
 } from '@aws-sdk/client-s3'
 import { getSignedUrl } from '@aws-sdk/s3-request-presigner'
 
+import type { AnyContextable } from '../../context'
 import type { StorageAdapter } from '../generic-adapter'
 
-export class StorageAdapterS3 implements StorageAdapter {
+export class StorageAdapterS3<TContext extends AnyContextable = AnyContextable>
+  implements StorageAdapter<TContext>
+{
   private AWSClient: S3Client
   private bucket: string
   public name = 'S3'
 
-  private constructor(options: { bucket: string; clientConfig: S3ClientConfig }) {
+  public readonly context: TContext
+  public imageBaseUrl: string
+
+  private constructor(options: {
+    bucket: string
+    clientConfig: S3ClientConfig
+    context: TContext
+    imageBaseUrl: string
+  }) {
     this.AWSClient = new S3Client(options.clientConfig)
     this.bucket = options.bucket
+    this.context = options.context
+    this.imageBaseUrl = options.imageBaseUrl
   }
 
   // Incase where you need to create the client automatically
-  public static initialize(options: { bucket: string; clientConfig: S3ClientConfig }) {
-    return new StorageAdapterS3(options)
+  public static initialize(
+    context: AnyContextable,
+    options: { bucket: string; clientConfig: S3ClientConfig; imageBaseUrl: string }
+  ) {
+    return new StorageAdapterS3({
+      context,
+      bucket: options.bucket,
+      clientConfig: options.clientConfig,
+      imageBaseUrl: options.imageBaseUrl,
+    })
   }
 
   /**
@@ -54,6 +76,23 @@ export class StorageAdapterS3 implements StorageAdapter {
     return {
       message: 'File upload signed URL request success',
       data: { putObjectUrl: uploadSignedUrl },
+    }
+  }
+
+  /**
+   * @description Get the signed URL for deleting the object from S3
+   */
+  public async generateDeleteObjectSignedUrl(arg: { key: string }) {
+    const deleteCommand = new DeleteObjectCommand({
+      Bucket: this.bucket,
+      Key: arg.key,
+    })
+
+    const deleteSignedUrl = await getSignedUrl(this.AWSClient, deleteCommand, { expiresIn: 3600 })
+
+    return {
+      message: 'File delete signed URL request success',
+      data: { deleteObjectUrl: deleteSignedUrl },
     }
   }
 }