+++ examples/erp/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @example/erp
 
+## 0.1.0-alpha.32
+
+### Patch Changes
+
+- [#154](https://github.com/softnetics/genseki/pull/154) [`4f04302`](https://github.com/softnetics/genseki/commit/4f04302a46e85a52b2f07156ec77fe55a17a4367) Thanks [@jettapat-metier](https://github.com/jettapat-metier)! - [Feature] Add delete object from storage
+
+- Updated dependencies [[`4f04302`](https://github.com/softnetics/genseki/commit/4f04302a46e85a52b2f07156ec77fe55a17a4367)]:
+  - @genseki/react@0.1.0-alpha.34
+  - @genseki/next@0.1.0-alpha.34
+  - @genseki/plugins@0.1.0-alpha.34
+  - @genseki/prisma-generator@0.1.0-alpha.34
+  - @genseki/react-query@0.1.0-alpha.34
+  - @genseki/rest@0.1.0-alpha.34
+
+## 0.1.0-alpha.31
+
+### Patch Changes
+
+- Updated dependencies [[`b1908d1`](https://github.com/softnetics/genseki/commit/b1908d164c623e2e9e409e7426e5948c75783e40)]:
+  - @genseki/react@0.1.0-alpha.33
+  - @genseki/next@0.1.0-alpha.33
+  - @genseki/plugins@0.1.0-alpha.33
+  - @genseki/prisma-generator@0.1.0-alpha.33
+  - @genseki/react-query@0.1.0-alpha.33
+  - @genseki/rest@0.1.0-alpha.33
+
 ## 0.1.0-alpha.30
 
 ### Patch Changes