+++ packages/react-query/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @genseki/react-query
 
+## 0.1.0-alpha.34
+
+## 0.1.0-alpha.33
+
 ## 0.1.0-alpha.32
 
 ## 0.1.0-alpha.31